pyproject.toml
[tool.poetry]
name = "rearq"
version = "0.2.7"
description = "Rewrite arq and make improvement."
authors = ["long2ice <long2ice@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/long2ice/rearq.git"
repository = "https://github.com/long2ice/rearq.git"
documentation = "https://github.com/long2ice/rearq/blob/master/README.md"
keywords = ["asyncio", "task", "arq", "queue", "distributed"]
packages = [
    { include = "rearq" }
]
include = ["LICENSE", "README.md", "CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "*"
click = "*"
redis = "^4.2.0rc1"
crontab = "*"
uvicorn = { version = "*", extras = ["standard"] }
fastapi = "*"
loguru = "*"
jinja2 = "*"
aiofiles = "*"
tortoise-orm = "^0.18.1"
asyncmy = { version = "*", optional = true }
asyncpg = { version = "*", optional = true }

[tool.poetry.dev-dependencies]
pyproject-flake8 = "*"
isort = "*"
black = "*"
pytest = "*"
pytest-xdist = "*"
bandit = "*"
pytest-asyncio = "*"
cryptography = "*"

[tool.poetry.extras]
postgres = ["asyncpg"]
mysql = ["asyncmy"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
rearq = "rearq.cli:main"

[tool.black]
line-length = 100
target-version = ['py36', 'py37', 'py38', 'py39']

[tool.pytest.ini_options]
asyncio_mode = 'auto'

[tool.mypy]
pretty = true
ignore_missing_imports = true

[tool.flake8]
ignore = 'E501,W503,E203'
